extract PlainCSV source class
Split Source::PlainCSV out into its own file
